Add: New CLI for downloading CPE match strings
Two new CLI tools are added to download CPE match strings from the NVD
API to either a consolidated JSON file or a PostgreSQL database.

These are also based on new generic and modular classes so parts like
the API downloader or the general structure of the JSON and database
writers can be easily reused.