Change: Reduce memory use of CPE match JSON writer
In the JSON write worker for CPE match strings, the response dict is
changed to camelCase keys in place instead of creating a copy in the
JsonEncoder.
The modified dict is also serialized to a char array with iterencode
instead of using json.dumps().

These changes reduce the memory usage when creating a JSON file with
a large number of match strings.