Add: exceptions constant for api access retries
We have seen problems with the application when the nvd API responds
properly http wise, but delivers json that cannot be decoded properly.

To keep retrying these API calls with the already existing retry
functionality until they succeed, which currently happens
after ~ 1-5 tries, we need to use a tuple with the already used
http error and also JSONDecodeError.

As this happens multiple times, we use a constant that can be used in
multiple places.

The constant is in a new constants file to avoid circular dependencies
from code that is called from the cli module back to the cli module
itself.

TODO:

1) We should refactor these accesses afterwards when the high prio for
   production fix is done, in order to have this functionality reusable
   in one place, because having the basically same stamina call in so
   many places calls for improvement.

2) also mode the existing constans from cli.py into wht constants file,
   also to avoid circular dependencies.